Moves componetent into appropriate folder/file & updated imports
