+++ src/App.js
@@ -1,13 +1,13 @@
 import "./App.css";
-import SnackShow from "./SnackShow";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import EditForm from "./EditForm";
-import NewForm from "./NewForm";
 
 // Pages
+import Edit from "./Pages/Edit";
 // import FourOFour from "./Pages/FourOFour";
 import Home from "./Pages/Home";
 import Index from "./Pages/Index";
+import NewSnack from "./Pages/NewSnack";
+import Show from "./Pages/Show";
 
 // Components
 import NavBar from "./Components/Navbar";
@@ -21,9 +21,9 @@ function App() {
           <Routes>
             <Route path="/" element={<Home />} />
             <Route path="/snacks" element={<Index />} />
-            <Route path="/snacks/:id" element={<SnackShow />} />
-            <Route path="/snacks/:id/edit" element={<EditForm />} />
-            <Route path="/snacks/new" element={<NewForm />} />
+            <Route path="/snacks/:id" element={<Show />} />
+            <Route path="/snacks/:id/edit" element={<Edit />} />
+            <Route path="/snacks/new" element={<NewSnack />} />
           </Routes>
         </main>
       </Router>
