--- src/Pages/New.js
@@ -1,12 +1,0 @@
-import SnackNewForm from "../Components/SnackNewForm";
-
-function New() {
-  return (
-    <div className="New">
-      <h2>New</h2>
-      <SnackNewForm />
-    </div>
-  );
-}
-
-export default New;
